+++ src/app/modules/lux-form/lux-file/lux-file-model/lux-file-object.interface.ts
@@ -3,6 +3,7 @@ import { Observable } from 'rxjs';
 export interface ILuxFileObject {
   name: string;
   type: string;
+  size?: number;
   content?: string | Blob;
   contentCallback?: Promise<any> | Observable<any> | any;
   namePrefix?: string;