+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Version 11.11.0
+## New
+- **lux-file-upload**: Neue File-Upload-Komponente [Issue xxx](https://github.com/IHK-GfI/lux-components/issues/xxx)
+
 ## Bug Fixes
 - **lux-stepper**: Es kann nicht über den Weiter-Button navigiert werden, wenn mehrere Schritte nachträglich hinzugefügt wurden. [Issue 167](https://github.com/IHK-GfI/lux-components/issues/167)
 - **lux-lookup-autocomplete**: Komponente wirft beim initialen Laden einen ExpressionChangedAfterItHasBeenCheckedError [Issue 165](https://github.com/IHK-GfI/lux-components/issues/165)